+++ src/AgOpenGPS/Forms/Settings/FormGraphHeading.cs
@@ -93,7 +93,7 @@ namespace AgOpenGPS
 
         private void FormSteerGraph_Load(object sender, EventArgs e)
         {
-            timer1.Interval = (int)((1 / (double)mf.gpsHz) * 1000);
+            timer1.Interval = (int)(1 / mf.gpsHz * 1000);
 
             if (!ScreenHelper.IsOnScreen(Bounds))
             {
